Add test for `dots add --help`
